requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
certifi==2024.8.30
    # via requests
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via depth-anything (pyproject.toml)
colorama==0.4.6
    # via
    #   click
    #   tqdm
filelock==3.16.1
    # via
    #   huggingface-hub
    #   torch
    #   transformers
fsspec==2024.10.0
    # via
    #   huggingface-hub
    #   torch
huggingface-hub==0.26.2
    # via
    #   tokenizers
    #   transformers
idna==3.10
    # via requests
jinja2==3.1.4
    # via torch
markupsafe==3.0.2
    # via jinja2
mpmath==1.3.0
    # via sympy
networkx==3.4.2
    # via torch
numpy==2.1.2
    # via
    #   depth-anything (pyproject.toml)
    #   opencv-python
    #   torchvision
    #   transformers
opencv-python==4.10.0.84
    # via depth-anything (pyproject.toml)
packaging==24.1
    # via
    #   huggingface-hub
    #   transformers
pillow==11.0.0
    # via
    #   depth-anything (pyproject.toml)
    #   torchvision
python-dotenv==1.0.1
    # via depth-anything (pyproject.toml)
pyyaml==6.0.2
    # via
    #   huggingface-hub
    #   transformers
regex==2024.9.11
    # via transformers
requests==2.32.3
    # via
    #   huggingface-hub
    #   transformers
safetensors==0.4.5
    # via transformers
sympy==1.13.1
    # via torch
tokenizers==0.20.1
    # via transformers
torch==2.5.1
    # via
    #   depth-anything (pyproject.toml)
    #   torchvision
torchvision==0.20.1
    # via depth-anything (pyproject.toml)
tqdm==4.66.6
    # via
    #   huggingface-hub
    #   transformers
transformers==4.46.1
    # via depth-anything (pyproject.toml)
typing-extensions==4.12.2
    # via
    #   huggingface-hub
    #   torch
urllib3==2.2.3
    # via requests
